Adding test cases for pure functional components
