+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-docgen-typescript",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "",
   "main": "lib/index.js",
   "scripts": {
@@ -16,6 +16,8 @@
     "typescript": "^2.2.1"
   },
   "devDependencies": {
+    "@types/chai": "^3.4.35",
+    "@types/mocha": "^2.2.39",
     "@types/node": "^7.0.5",
     "@types/react": "^15.0.14",
     "@types/react-dom": "^0.14.23",
@@ -24,11 +26,12 @@
     "babel-loader": "^6.2.4",
     "chai": "^3.5.0",
     "file-loader": "^0.8.5",
-    "mocha": "^2.4.5",
+    "mocha": "^2.5.3",
     "react": "^15.4.2",
     "react-bluekit": "^0.4.1",
     "react-dom": "^15.4.2",
-    "react-styleguidist": "^2.2.1"
+    "react-styleguidist": "^2.2.1",
+    "ts-node": "^2.1.0"
   },
   "files": [
     "lib",
