+++ src/__tests__/data/Row.tsx
@@ -1,0 +1,36 @@
+import * as React from 'react';
+
+/**
+ * Row properties.
+ */
+export interface IRowProps {
+    /** prop1 description */
+    prop1?: string;
+    /** prop2 description */
+    prop2: number;
+    /**
+     * prop3 description
+     */
+    prop3: () => void;
+    /** prop4 description */
+    prop4: 'option1' | 'option2' | "option3";
+}
+
+/**
+ * Form row.
+ */
+export const Row = (props: IRowProps) => {
+    const innerFunc = (props: IRowProps) => {
+        return <span>Inner Func</span>
+    };
+    const innerNonExportedFunc = (props: IRowProps) => {
+        return <span>Inner Func</span>
+    };
+    return <div>Test</div>;
+};
+
+const nonExportedFunc = (props: IRowProps) => {
+    return <div>No Export</div>
+};
+
+export default Row;
